Rename schema modules and update imports
