+++ app/routers/notifications.py
@@ -9,7 +9,7 @@ from app.core.database import get_db
 from app.models.notification import Notification
 from app.models.user import User
 from app.routers.users import get_current_user
-from app.schemas.notification import NotificationOut  # <-- import BEFORE usage
+from app.schemas.notification_schema import NotificationOut  # <-- import BEFORE usage
 
 router = APIRouter(prefix="/notifications", tags=["notifications"])
