+++ app/routers/tasks.py
@@ -6,7 +6,7 @@ from sqlalchemy import select, and_, or_
 from app.core.database import get_db
 from app.models.task import Task, Priority, Status
 from app.models.user import User, Role
-from app.schemas.task import TaskCreate, TaskUpdate, TaskOut
+from app.schemas.task_schema import TaskCreate, TaskUpdate, TaskOut
 from app.routers.users import get_current_user
 
 router = APIRouter(prefix="/tasks", tags=["tasks"])
